# CPU: native attention fallback + legacy KVCache path to enable sglang_jax launch