--- a/python/sgl_jax/srt/utils/jax_utils.py
+++ b/python/sgl_jax/srt/utils/jax_utils.py
@@ -97,3 +97,16 @@ def device_array(mesh, *data, sharding=None, **kwargs) -> jax.Array:
     if sharding is None:
         sharding = NamedSharding(mesh, PartitionSpec())
     return jax.device_put(*data, device=sharding, **kwargs)
+
+
+def is_tpu_runtime() -> bool:
+    """Return True if the current JAX runtime is on TPU devices.
+
+    Prefer checking actual devices; fall back to default backend if necessary.
+    """
+    try:
+        devs = jax.devices()
+        return len(devs) > 0 and all(d.platform == "tpu" for d in devs)
+    except Exception:
+        return jax.default_backend() == "tpu"
+